+++ lib/pueri.rb
@@ -2,6 +2,7 @@
 
 require 'pueri/age'
 require 'pueri/dosecalc'
+require 'pueri/dosecheck'
 require 'pueri/neuro'
 require 'pueri/vax'
 require 'pueri/version'